+++ packages/frameworks/angular-core/src/pagination/core-pagination-page-metadata.directive.ts
@@ -11,7 +11,7 @@ import {usePaginationContext} from "./pagination-context.service"
 export class CorePaginationPageMetadataDirective implements OnInit {
   readonly paginationContext = usePaginationContext()
 
-  private readonly trackBindings = useTrackBindings(() =>
+  protected readonly trackBindings = useTrackBindings(() =>
     this.paginationContext().getPageMetadataBindings(),
   )