+++ packages/frameworks/angular/src/pagination/qds-pagination-context.service.ts
@@ -1,0 +1,23 @@
+// Copyright (c) Qualcomm Technologies, Inc. and/or its subsidiaries.
+// SPDX-License-Identifier: BSD-3-Clause-Clear
+
+import {Injectable} from "@angular/core"
+
+import {
+  type ApiContext,
+  BaseApiContextService,
+  createApiContext,
+} from "@qualcomm-ui/angular-core/machine"
+import type {QdsPaginationApi} from "@qualcomm-ui/qds-core/pagination"
+
+@Injectable()
+export class QdsPaginationContextService extends BaseApiContextService<QdsPaginationApi> {}
+
+export const [
+  QDS_PAGINATION_CONTEXT,
+  useQdsPaginationContext,
+  provideQdsPaginationContext,
+]: ApiContext<QdsPaginationApi> = createApiContext<QdsPaginationApi>(
+  "QdsPaginationContext",
+  QdsPaginationContextService,
+)