+++ packages/common/qds-core/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @qualcomm-ui/qds-core
 
+## 1.4.0 (2025/11/18)
+
+### Features
+
+- add action group api
+
+### Bug Fixes
+
+- [icon]: set size css variable when using custom size
+- [button]: text-decoration defaults to inherit
+
+### Miscellaneous Chores
+
+- **deps:** update dependencies [@qualcomm-ui/core]
+
 ## 1.3.0 (2025/11/17)
 
 ### Features